rename `files` to `appDir`
