+++ src/constants.js
@@ -1,0 +1,14 @@
+export const Options = {
+  mode: "run", // run or build
+  // Run API
+  appDir: ".",
+  caheDir: "./cache",
+  version: "0.67.1",
+  flavour: "sdk",
+  platform: "",
+  mirror: "https://nwjs.io/",
+  downloadUrl: "https://dl.nwjs.io",
+  // Build API
+  buildDir: "./build",
+  forceDownload: false,
+};
