+++ src/getArchitecture.js
