+++ src/getPlatform.js
