+++ src/parseOptions.js
